[explorer][delegation] add wallet APT balance to add stake flow
